move copy_data
